Rename parser as template
